Updated networking to reflect map building changes.

+++ src/main/java/com/forgestorm/server/network/game/packet/in/WorldBuilderPacketIn.java
@@ -13,7 +13,7 @@ public class WorldBuilderPacketIn implements PacketListener<WorldBuilderPacketIn
     @Override
     public PacketData decodePacket(ClientHandler clientHandler) {
         LayerDefinition layerDefinition = LayerDefinition.getLayerDefinition(clientHandler.readByte());
-        short textureId = clientHandler.readShort();
+        int textureId = clientHandler.readInt();
         short tileX = clientHandler.readShort();
         short tileY = clientHandler.readShort();
         return new WorldBuilderPacket(clientHandler.getPlayer(), layerDefinition, textureId, tileX, tileY);
@@ -39,7 +39,7 @@ public class WorldBuilderPacketIn implements PacketListener<WorldBuilderPacketIn
     class WorldBuilderPacket extends PacketData {
         private Player playerSender;
         private LayerDefinition layerDefinition;
-        private short textureId;
+        private int textureId;
         private short tileX;
         private short tileY;
     }
